use INTEGER instead of SMALLINT

+++ src/main/resources/db/migration/V4__book_fields.sql
@@ -2,7 +2,7 @@ ALTER TABLE books
 ADD COLUMN key VARCHAR(255),
 ADD COLUMN subtitle VARCHAR(255),
 ADD COLUMN isbn10 TEXT[],
-ADD COLUMN page_number SMALLINT,
+ADD COLUMN page_number INTEGER,
 ADD COLUMN publish_date VARCHAR(255),
 ADD COLUMN publish_country VARCHAR(255),
 ADD COLUMN publish_places TEXT[],
